lightning-mock: refactor tests
